Thread the expression through the forms
As the parser begins to take shape, there are going to be a number of
different sub-parsers that will be used on lines.

Rather than using nested function calls, provide a simpler call
hierarchy by using the -> (arrow) form.

This may need to be switched, depending on whether the 'second' param,
or 'last' param is used.